backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
gunicorn==23.0.0
openai==1.68.0
requests==2.31.0
google-cloud-vision==3.10.1
google-cloud-storage==3.1.0
opencv-python>=4.9.0.80
pytesseract==0.3.10
numpy==1.26.4
waitress==3.0.2
psycopg2-binary==2.9.9
firebase-admin==6.7.0
